+++ changelog.md
@@ -1,13 +1,19 @@
 # Luma Tools Changelog
 
+## Version 0.4.1.23
+- Always append input filename to ComfyUI output files
+
+## Version 0.4.1.22
+- Feature request improvements
+
 ## Version 0.4.1.21
-Fix 3d viewer
+- Fix 3d viewer
 
 ## Version 0.4.1.20
-- move user requests to user folders
+- Move user requests to user folders
 
 ## Version 0.4.1.19
-Support input images from custom nodes
+- Support input images from custom nodes
 
 ## Version 0.4.1.18
 - Fix input file handling edge cases, always look in their hardcoded default input directory.